requirements.txt
Django>=5.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
django-environ>=0.11
gunicorn>=21.2
djangorestframework>=3.14
drf-spectacular>=0.27
